feat(doc) add CovWriter that generate coverage file
Relate #1149
