+++ lib/theme/nico.js
@@ -41,7 +41,8 @@ exports.writers = [
   'nico.PageWriter',
   'nico.StaticWriter',
   'nico.FileWriter',
-  'nico.MochaWriter'
+  'nico.MochaWriter',
+  'nico.CovWriter'
 ];
 
 exports.middlewares = [
