+++ lib/theme/theme.js
@@ -4,14 +4,18 @@ module.exports = function(nico) {
 
   var exports = {};
 
+  var Instrumenter = require('istanbul').Instrumenter;
+  var instrumenter = new Instrumenter();
   var ReactTools = require('react-tools');
   var path = require('path');
+  var fs = require('fs');
   var util = require('util');
   var file = nico.sdk.file;
   var glob = nico.sdk.file.glob;
   var BaseWriter = nico.BaseWriter;
   var getDependencies = require('../sdk/module').getDependencies;
   var getSourceFiles = require('../sdk/module').getSourceFiles;
+  var mo = require('../sdk/module');
   var crequire = require('crequire');
 
   function MochaWriter() {}
@@ -26,6 +30,22 @@ module.exports = function(nico) {
     });
   };
   nico.MochaWriter = MochaWriter;
+
+  function CovWriter() {}
+  util.inherits(CovWriter, BaseWriter);
+  CovWriter.prototype.run = function() {
+    var option = nico.sdk.option;
+    var dest = option.get('outputdir');
+    mo.getSourceFiles().forEach(function(item) {
+      if (!path.extname(item)) {
+        item = item + '.js';
+      }
+      if (path.extname(item) === '.js') {
+        processCovFile(item, dest);
+      }
+    });
+  };
+  nico.CovWriter = CovWriter;
 
   exports.name = 'spmjs';
   exports.version = '1.0';
@@ -224,5 +244,20 @@ module.exports = function(nico) {
   }
 
   return exports;
+
+  function processCovFile(source, destDir) {
+    var dest;
+    if (/\.\w+$/.test(source)) {
+      dest = source.replace(/(\.\w+)$/, '-cov$1');
+    } else {
+      dest = source + '-cov';
+    }
+
+    var file = path.join(process.cwd(), source);
+    var code = fs.readFileSync(file).toString();
+    var changed = instrumenter.instrumentSync(code, file);
+    fs.writeFileSync(path.join(destDir, dest), changed);
+  }
 };
+
